backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.4.2
pyyaml==6.0.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 